Move NoData composable into "designsystem" package.
